requirements.txt
# Bioiain dependencies #################################################
# Should match pyproject.toml


# Critical ############################################################
#python == 3.14
numpy == 1.26.4
biopython == 1.86
pandas == 3.0.0
matplotlib == 3.10.8
scikit-learn == 1.8.0
scipy == 1.17.0
#pymol-open-source == 2.5.0#conda-forge::pymol-open-source

# Essential ###########################################################
setuptools >= 80.10.1
typing-extensions >= 4.15.0
requests >= 2.32.5
unidecode >= 1.4.0
xmltodict >= 0.14.2

# ML ###################################################################
torch >= 2.10.0
torchvision >= 0.25.0
transformers >= 5.0.0
# foldseek==10.941cd33 #bioconda::foldseek



# Dev ##################################################################

#dev

pdoc3
twine
build
